create image delete modal

+++ public/js/product.js
@@ -111,15 +111,24 @@ function restoreSpecification(productId, specificationId) {
 }
 
 function deleteImage(productId, imageId) {
-    $.ajax({
-        type: 'DELETE',
-        url: `/products/${productId}/images/${imageId}`,
-        success: (response) => {
-            $(`img[data-image-id=${imageId}]`).remove();
-        },
-        error: (error) => {
-            console.error(error);
-        }
+    Swal.fire({
+        title: 'Are you sure you want to delete?',
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonText: 'Delete'
+    }).then(() => {
+        $.ajax({
+            type: 'DELETE',
+            url: `/products/${productId}/images/${imageId}`,
+            success: (response) => {
+                Swal.fire('Deleted!', '', 'success');
+
+                $(`div[data-image-id=${imageId}]`).remove();
+            },
+            error: (error) => {
+                console.error(error);
+            }
+        });
     });
 }
